Path: Refactor `Extensions` feature code into independent modules
Move `Extensions` code into independent modules to allow for easier application to other operations in Path workbench.
The only changes to the `Extensions` code base is to update references to same functions in new modules.
delete gui


cleanup


remove
